Fixes now playing progress bar creation

+++ src/main/java/com/reign/kat/lib/voice/newvoice/RequestedTrack.java
@@ -38,7 +38,7 @@ public class RequestedTrack
      */
     public float getPercentComplete()
     {
-        return (float)(track.getPosition() / track.getDuration());
+        return (float)track.getPosition() / (float)track.getDuration();
     }
 
     public String getPositionAsTimestamp()
